review changes and moving all configs to extension
